% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voro.R
\docType{package}
\name{voro-package}
\alias{voro}
\alias{voro-package}
\title{voro: R interface to the Voro++ software library}
\description{
Voro++ (http://math.lbl.gov/voro++) is an open source software library for the computation of the Voronoi diagram, a widely-used tessellation that has applications in many scientific fields. This package wraps the library and provides an R interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nevrome/voro}
  \item Report bugs at \url{https://github.com/nevrome/voro}
}

}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID})

}
\keyword{internal}
